refactor join predicate pushdown logic to be used in mv rewrite (#23413)
