+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/PushDownJoinOnClauseRule.java
@@ -17,18 +17,17 @@ package com.starrocks.sql.optimizer.rule.transformation;
 import com.google.common.collect.Lists;
 import com.starrocks.sql.optimizer.OptExpression;
 import com.starrocks.sql.optimizer.OptimizerContext;
-import com.starrocks.sql.optimizer.Utils;
-import com.starrocks.sql.optimizer.base.ColumnRefSet;
 import com.starrocks.sql.optimizer.operator.OperatorType;
 import com.starrocks.sql.optimizer.operator.logical.LogicalJoinOperator;
 import com.starrocks.sql.optimizer.operator.pattern.Pattern;
 import com.starrocks.sql.optimizer.operator.scalar.ScalarOperator;
+import com.starrocks.sql.optimizer.rewrite.JoinPredicatePushdown;
 import com.starrocks.sql.optimizer.rule.RuleType;
 
 import java.util.Collections;
 import java.util.List;
 
-public class PushDownJoinOnClauseRule extends PushDownJoinPredicateBase {
+public class PushDownJoinOnClauseRule extends TransformationRule {
     public PushDownJoinOnClauseRule() {
         super(RuleType.TF_PUSH_DOWN_JOIN_CLAUSE, Pattern.create(OperatorType.LOGICAL_JOIN).
                 addChildren(Pattern.create(OperatorType.PATTERN_LEAF), Pattern.create(OperatorType.PATTERN_LEAF)));
@@ -48,51 +47,15 @@ public class PushDownJoinOnClauseRule extends PushDownJoinPredicateBase {
     public List<OptExpression> transform(OptExpression input, OptimizerContext context) {
         List<OptExpression> children = Lists.newArrayList(input.getInputs());
         LogicalJoinOperator join = (LogicalJoinOperator) input.getOp();
-
         ScalarOperator on = join.getOnPredicate();
-
-        on = rangePredicateDerive(on);
-        on = equivalenceDeriveOnPredicate(on, input, join);
-
-        OptExpression root = pushDownOnPredicate(input, on);
+        JoinPredicatePushdown joinPredicatePushdown = new JoinPredicatePushdown(
+                input, true, false, context.getColumnRefFactory());
+        OptExpression root = joinPredicatePushdown.pushdown(join.getOnPredicate());
         ((LogicalJoinOperator) root.getOp()).setHasPushDownJoinOnClause(true);
         if (root.getOp().equals(input.getOp()) && on.equals(join.getOnPredicate()) &&
                 children.equals(root.getInputs())) {
             return Collections.emptyList();
         }
         return Lists.newArrayList(root);
-    }
-
-    ScalarOperator equivalenceDeriveOnPredicate(ScalarOperator on, OptExpression joinOpt, LogicalJoinOperator join) {
-        // For SQl: select * from t1 left join t2 on t1.id = t2.id and t1.id > 1
-        // Infer t2.id > 1 and Push down it to right child
-        if (!join.getJoinType().isInnerJoin() && !join.getJoinType().isSemiJoin() &&
-                !join.getJoinType().isOuterJoin()) {
-            return on;
-        }
-
-        List<ScalarOperator> pushDown = Lists.newArrayList(on);
-        ColumnRefSet leftOutputColumns = joinOpt.getInputs().get(0).getOutputColumns();
-        ColumnRefSet rightOutputColumns = joinOpt.getInputs().get(1).getOutputColumns();
-
-        ScalarOperator derivedPredicate = equivalenceDerive(on, false);
-        List<ScalarOperator> derivedPredicates = Utils.extractConjuncts(derivedPredicate);
-
-        if (join.getJoinType().isInnerJoin() || join.getJoinType().isSemiJoin()) {
-            return Utils.compoundAnd(on, derivedPredicate);
-        } else if (join.getJoinType().isLeftOuterJoin()) {
-            for (ScalarOperator p : derivedPredicates) {
-                if (rightOutputColumns.containsAll(p.getUsedColumns())) {
-                    pushDown.add(p);
-                }
-            }
-        } else if (join.getJoinType().isRightOuterJoin()) {
-            for (ScalarOperator p : derivedPredicates) {
-                if (leftOutputColumns.containsAll(p.getUsedColumns())) {
-                    pushDown.add(p);
-                }
-            }
-        }
-        return Utils.compoundAnd(pushDown);
     }
 }
